refactor: move & rename bus events helper function
