+++ apps/omg_watcher/test/omg_watcher/exit_processor/finalizations_test.exs
@@ -329,53 +329,4 @@ defmodule OMG.Watcher.ExitProcessor.FinalizationsTest do
         Core.finalize_in_flight_exits(processor, [finalization1, finalization2], %{})
     end
   end
-
-  describe "to_bus_events/1" do
-    setup _ do
-      {:ok,
-       %{
-         finalizations: [
-           %{in_flight_exit_id: <<1::192>>, log_index: 1, root_chain_txhash: <<1::256>>},
-           %{in_flight_exit_id: <<2::192>>, log_index: 2, root_chain_txhash: <<2::256>>}
-         ],
-         utxos: [
-           Utxo.position(1, 0, 0),
-           Utxo.position(1000, 0, 0),
-           Utxo.position(2000, 0, 0)
-         ]
-       }}
-    end
-
-    test "mapping single finalization", %{finalizations: [f1 | _], utxos: [utxo_1 | _]} do
-      utxo_pos = Utxo.Position.encode(utxo_1)
-
-      assert [
-               %{log_index: 1, root_chain_txhash: <<1::256>>, call_data: %{utxo_pos: ^utxo_pos}}
-             ] = Core.to_bus_events([{f1, [utxo_1]}])
-    end
-
-    test "mapping multiple finalizations", %{finalizations: [f1, f2 | _], utxos: utxos} do
-      [utxo_1, utxo_2, utxo_3 | _] = utxos
-      [utxo_pos_1, utxo_pos_2, utxo_pos_3 | _] = Enum.map(utxos, &Utxo.Position.encode/1)
-
-      assert [
-               %{log_index: 2, root_chain_txhash: <<2::256>>, call_data: %{utxo_pos: ^utxo_pos_2}},
-               %{log_index: 2, root_chain_txhash: <<2::256>>, call_data: %{utxo_pos: ^utxo_pos_3}},
-               %{log_index: 1, root_chain_txhash: <<1::256>>, call_data: %{utxo_pos: ^utxo_pos_1}}
-             ] = Core.to_bus_events([{f1, [utxo_1]}, {f2, [utxo_2, utxo_3]}])
-    end
-
-    test "finalization without exiting utxos does not produce events",
-         %{finalizations: [f1, f2 | _], utxos: [utxo_1 | _]} do
-      utxo_pos = Utxo.Position.encode(utxo_1)
-
-      assert [
-               %{log_index: 2, root_chain_txhash: <<2::256>>, call_data: %{utxo_pos: ^utxo_pos}}
-             ] = Core.to_bus_events([{f1, []}, {f2, [utxo_1]}])
-    end
-
-    test "empty finalization list does not produce events" do
-      assert [] = Core.to_bus_events([])
-    end
-  end
 end
